+++ CHANGELOG.md
@@ -8,7 +8,11 @@
 
 ## [0.0.0+2] - Adds image to documentation
 
-<img src="./sample1.png"/><br/>
-* Other documentation improvements
+Other documentation improvements
+<img src="https://github.com/playlinesdev/pl_grid/blob/master/sample1.png?raw=true"/>
 
 ## [0.0.0+4] - Adds MIT License info
+
+## [0.0.1+0] - Adds lots of customization possibilities to have control of each cell
+Makes it also possible to render Widgets on any cell
+<img src="https://github.com/playlinesdev/pl_grid/blob/master/sample4.png?raw=true"/>